Update p2p module with good structure
